net: hns3: fix vport->bw_limit overflow problem
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

The next-net has fixed it, so do not need upstreaming.

+++ drivers/net/ethernet/hisilicon/hns3/hns3pf/hclge_main.h
@@ -870,7 +870,7 @@ struct hclge_vport {
 	u16 alloc_rss_size;
 
 	u16 qs_offset;
-	u16 bw_limit;		/* VSI BW Limit (0 = disabled) */
+	u32 bw_limit;		/* VSI BW Limit (0 = disabled) */
 	u8  dwrr;
 
 	struct list_head vlan_list;     /* Store VF vlan table */
